Attach 2 PDFs

+++ app/Http/Controllers/EmailController.php
@@ -20,14 +20,15 @@ class EmailController extends Controller
         $count = 0;
         foreach ($emails as $item) {
             $chk_send = $this->send_email_notification($item->id);
-            if(!$chk_send['success']){
+
+            if($chk_send['success'] == false){
                 flash($chk_send['message'])->error();
                 return redirect(route('app.email.show',$tmail_id));
             }
             $count++;
         }
         flash($count . ' correos enviados')->success();
-        return redirect(route('app.email.show',$tmail_id));
+        return redirect(route('app.email.show', $tmail_id));
         // return ['success'=>true];
     }
 
@@ -39,10 +40,10 @@ class EmailController extends Controller
         $wdocente = $toUser->wdocente;
         $cdocente = $toUser->cdocente;
 
-        $file_to_attach = storage_path() . DIRECTORY_SEPARATOR . 'reports'. DIRECTORY_SEPARATOR .'report_' . $toUser->cdocente . '.pdf';
-        $file_name = 'report_' . $toUser->wdocente . '.pdf';    
+        // $file_to_attach = storage_path() . DIRECTORY_SEPARATOR . 'reports'. DIRECTORY_SEPARATOR .'report_' . $toUser->cdocente . '.pdf';
+        // $file_name = 'report_' . $toUser->wdocente . '.pdf';    
 
-        if(file_exists($file_to_attach)){
+        // if(file_exists($file_to_attach)){
             $data = [
                 'from' => $email->from,
                 'dfrom' => 'Departamento Académico FCEC',
@@ -55,8 +56,8 @@ class EmailController extends Controller
             ];
 
             $attach = [
-                'file_to_attach' => $file_to_attach,
-                'file_name' => $file_name,            
+                ['file_to_attach' => $email->file_to_attach1, 'file_name' => $email->file_name1 ],
+                ['file_to_attach' => $email->file_to_attach2, 'file_name' => $email->file_name2 ]
             ];
 
             try {
@@ -64,27 +65,29 @@ class EmailController extends Controller
                         $message->to($data['to'], '');
                         $message->from('ucss.fcec.lim@gmail.com', 'UCSS – FCEC');
                         $message->subject($data['subject']);
-                        $message->attach( $attach['file_to_attach'], 
-                            [ 'as'=>$attach['file_name'], 'mime' => 'application/pdf']);
+                        foreach ($attach as $item) {
+                            $message->attach( $item['file_to_attach'], 
+                                [ 'as'=>$item['file_name'], 'mime' => 'application/pdf']);
+                        }
                     });         
 
                     $email->send_date = now();
                     $email->save();
 
-                    // flash('Email enviado a: ' . $data['to'])->success();
-                    // return response()->json(['success' => true ]);
             } catch (Exception $e) {
+                return ['success'=>false, 'message' => 'Error al enviar el correo electrónico.'];
+
                 dd('Error de proceso send_email_notification.');
                 // flash('Error de proceso send_email_notification.')->error();
                 // return back();
             }
-            return true;
-        }else{
-            // flash('Genere el archivo PDF de: '.$wdocente)->error();
-            // return back();
-            // return redirect(route('app.schedule.index'));
-            return ['success'=>false, 'message' => 'Genere el archivo PDF de: '.$wdocente];
-        }
+            return ['success'=>true];
+        // }else{
+        //     // flash('Genere el archivo PDF de: '.$wdocente)->error();
+        //     // return back();
+        //     // return redirect(route('app.schedule.index'));
+        //     return ['success'=>false, 'message' => 'Genere el archivo PDF de: '.$wdocente];
+        // }
 
     }   // end of send_notification()
 
@@ -136,44 +139,55 @@ class EmailController extends Controller
     public function store(Request $request)
     {
         $tmail = Tmail::findOrFail($request->tmail_id);
-        $users = [];
-        foreach ($request->chk as $key => $value) {
-            $user = User::findOrFail($key);
-            array_push($users, $user);
-        }
-        try {
-            $olds = Email::where('tmail_id', $tmail->id)
-                        ->whereNull('send_date')->get();
-            foreach ($olds as $old) {
-                if(!array_key_exists($old->user_id_to, $request->chk)){
-                    $old->delete();
-                }
+        if(is_null($request->chk)){
+            flash('No se ha seleccionado nuevos envíos de correo.')->error();
+            return redirect(route('app.email.index', $tmail->id));
+        }else{
+            $users = [];
+
+            foreach ($request->chk as $key => $value) {
+                $user = User::findOrFail($key);
+                array_push($users, $user);
             }
-            foreach ($users as $user) {
-                $old = Email::where('tmail_id', $tmail->id)
-                        ->where('to', $user->email)->get();
-                if($old->count() == 0){
-                    $responseR = Report::reportDownload_storage($user->id);
-                    $responseC = Report::cronoDownload_storage($user->id);
-                    if($responseR['success'] && $responseC['success']){
-                        $email = new Email;
-                        $email->tmail_id = $tmail->id ;
-                        $email->from = env('MAIL_USERNAME');
-                        $email->user_id_to = $user->id ;
-                        $email->to = $user->email ;
-                        $email->view = $tmail->view ;
-                        $email->limit_date = $tmail->limit_date;
-                        $email->save();
-                    } else {
-                        flash('Error al generar PDF: '.$user->name)->error();
-                        return redirect(route('app.email.index', $tmail->id));
-                        break;
+            try {
+                $olds = Email::where('tmail_id', $tmail->id)
+                            ->whereNull('send_date')->get();
+                foreach ($olds as $old) {
+                    if(!array_key_exists($old->user_id_to, $request->chk)){
+                        $old->delete();
                     }
                 }
+                foreach ($users as $user) {
+                    $old = Email::where('tmail_id', $tmail->id)
+                            ->where('to', $user->email)->get();
+                    if($old->count() == 0){
+                        $responseR = Report::reportDownload_storage($user->id);
+                        $responseC = Report::cronoDownload_storage($user->id);
+                        if($responseR['success'] && $responseC['success']){
+                            $email = new Email;
+                            $email->tmail_id = $tmail->id ;
+                            $email->from = env('MAIL_USERNAME');
+                            $email->user_id_to = $user->id ;
+                            $email->to = $user->email ;
+                            $email->view = $tmail->view ;
+                            $email->limit_date = $tmail->limit_date;
+                            $email->file_to_attach1 = $responseR['file_to_attach'];
+                            $email->file_name1 = $responseR['file_name'];
+                            $email->file_to_attach2 = $responseC['file_to_attach'];
+                            $email->file_name2 = $responseC['file_name'];
+                            $email->save();
+                        } else {
+                            flash('Error al generar PDF: '.$user->name)->error();
+                            return redirect(route('app.email.index', $tmail->id));
+                            break;
+                        }
+                    }
+                }
+                return redirect(route('app.email.show', $tmail->id));
+            } catch (Exception $e) {
+                dd('error EmailController@store', e);
             }
-            return redirect(route('app.email.show', $tmail->id));
-        } catch (Exception $e) {
-            dd('error EmailController@store', e);
+
         }
     }
 
@@ -188,13 +202,18 @@ class EmailController extends Controller
             $user->setAttribute('reply_date', $item->reply_date);
             array_push($docs, $user);
         }
+        $chk_send = [];
+        $chk_send = array_filter($docs, function($i)
+        {
+            return ( $i['send_date'] == null ) ? true : false;
+        });
 
         usort($docs, function ($item1, $item2) {
             return $item1['name'] <=> $item2['name'];
         });
 
         return view('app.mail.email.show')
-            ->with('data', ['users'=>$docs, 'tmail'=> $tmail]);
+            ->with('data', ['users'=>$docs, 'tmail'=> $tmail, 'chk_send' => count($chk_send)]);
     }
 
     public function edit($id)
